update Kendall's my_solution.js
What is 'this' referring to? Think programming-wise rather than in the terms of the game.
'This' apparently used to only refer to an object but with recent updates it can have any value specifically in functions. It is never static.
What does the () do in JavaScript?
They call the function and can also add parameters to function brackets. They are not always empty but always necessary.
What is the point of the semicolons?
Semicolons in JavaScript are used to separate statements. Sometimes they are optional but always a good habit. They are required when two statements are on one line. You don't use semicolons after curly brackets (of course there are exceptions such as assignment statements).